# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Rent Management API
  version: 1.0.0

servers:
  - url: https://my-first-backend-uen5.onrender.com
    description: Production server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Users
    description: User management endpoints
  - name: Tasks
    description: Task management endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
        401:
          description: Invalid password
        404:
          description: User not found

  /users:
    get:
      summary: Get all users
      tags: [Users]
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    first_name:
                      type: string
                      example: John
                    middle_name:
                      type: string
                      example: Michael
                    last_name:
                      type: string
                      example: Doe
                    email:
                      type: string
                      example: "example@gmail.com"
                    phone_number:
                      type: string
                      example: "+2348012345678"
                    personal_house_address:
                      type: string
                      example: "123 Main Street, Lagos"
                    user_image:
                      type: string
                      example: "base64encodedstring"
                    verified:
                      type: boolean
                      example: true
                    user_type:
                      type: string
                      example: "tenant"

    post:
      summary: Register a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - user_type
              properties:
                email:
                  type: string
                password:
                  type: string
                user_type:
                  type: string
                  enum: [landlord, tenant]
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created. OTP sent (expires in 5 mins)."
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: "example@example.com"
                      user_type:
                        type: string
                        example: "tenant"

  /users/me:
    get:
      summary: Get current user details
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  first_name:
                    type: string
                    example: John
                  middle_name:
                    type: string
                    example: Michael
                  last_name:
                    type: string
                    example: Doe
                  email:
                    type: string
                    example: "example@gmail.com"
                  phone_number:
                    type: string
                    example: "+2348012345678"
                  personal_house_address:
                    type: string
                    example: "123 Main Street, Lagos"
                  user_image:
                    type: string
                    example: "base64encodedstring"
                  verified:
                    type: boolean
                    example: true
                  user_type:
                    type: string
                    example: "tenant"
        401:
          description: Unauthorized

    patch:
      summary: Update current user details
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  example: John
                middle_name:
                  type: string
                  example: Michael
                last_name:
                  type: string
                  example: Doe
                phone_number:
                  type: string
                  example: "+2348012345678"
                personal_house_address:
                  type: string
                  example: "123 Main Street, Lagos"
                user_image:
                  type: string
                  format: binary
                  example: "base64encodedstring"
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      first_name:
                        type: string
                        example: John
                      middle_name:
                        type: string
                        example: Michael
                      last_name:
                        type: string
                        example: Doe
                      phone_number:
                        type: string
                        example: "+2348012345678"
                      personal_house_address:
                        type: string
                        example: "123 Main Street, Lagos"
                      user_image:
                        type: string
                        example: "base64encodedstring"
                      verified:
                        type: boolean
                        example: true
                      user_type:
                        type: string
                        example: "tenant"
        400:
          description: No fields provided for update
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No fields to update"
        401:
          description: Unauthorized
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error updating user"
                  error:
                    type: string
                    example: "Database error message"

  /auth/verify-otp:
    post:
      summary: Verify OTP
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                otp:
                  type: string
      responses:
        200:
          description: Verified
        400:
          description: Invalid OTP

  /auth/resend-otp:
    post:
      summary: Resend OTP
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        200:
          description: OTP sent
        400:
          description: Already verified
        404:
          description: User not found

  /tasks:
    get:
      summary: Get all tasks
      tags: [Tasks]
      responses:
        200:
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    title:
                      type: string
                      example: "Fix the leak"
                    description:
                      type: string
                      example: "Leaking pipe in kitchen"
                    status:
                      type: string
                      example: "in progress"
                    created_at:
                      type: string
                      format: date-time
                      example: "2025-12-01T10:00:00Z"
                    updated_at:
                      type: string
                      format: date-time
                      example: "2025-12-01T11:00:00Z"